move static classes to their own files
