+++ libs/java/server_aws_common/src/test/java/io/athenz/server/aws/common/store/impl/AWSObjectStoreFactoryTest.java
@@ -26,8 +26,9 @@ import org.testng.annotations.Test;
 
 import software.amazon.awssdk.regions.Region;
 
+import java.sql.DriverManager;
+import java.sql.SQLException;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
@@ -158,44 +159,6 @@ public class AWSObjectStoreFactoryTest {
         assertNull(factory.getAuthTokenFromCandidateRegions(null, 3306, "rds-user"));
     }
 
-    static class CandidateRegionTestFactory extends AWSObjectStoreFactory {
-
-        private Region workingRegion;
-        private Region noTokenRegion;
-        final List<Region> regionsTried = new java.util.concurrent.CopyOnWriteArrayList<>();
-
-        CandidateRegionTestFactory(Region workingRegion) {
-            this.workingRegion = workingRegion;
-        }
-
-        void setWorkingRegion(Region workingRegion) {
-            this.workingRegion = workingRegion;
-        }
-
-        void setNoTokenRegion(Region noTokenRegion) {
-            this.noTokenRegion = noTokenRegion;
-        }
-
-        @Override
-        Region getRegion() {
-            return Region.US_EAST_1;
-        }
-
-        @Override
-        String getAuthToken(String hostname, int port, String rdsUser, Region region) {
-            regionsTried.add(region);
-            if (region.equals(noTokenRegion)) {
-                return null;
-            }
-            return "token-" + region.id();
-        }
-
-        @Override
-        boolean verifyConnection(String jdbcUrl, String rdsUser, String token) {
-            return token.equals("token-" + workingRegion.id());
-        }
-    }
-
     @Test
     public void testGetAuthTokenFromCandidateRegionsSkipsRegionWithNoToken() {
 
@@ -235,80 +198,19 @@ public class AWSObjectStoreFactoryTest {
         assertFalse(factory.verifyConnection("not-a-valid-jdbc-url", "user", "token"));
     }
 
-    /**
-     * Minimal fake JDBC driver so tests can exercise verifyConnection()'s success path
-     * without a live database. Only connect()/acceptsURL() matter; the returned Connection
-     * is a no-op proxy since verifyConnection() only checks it for non-null before closing it.
-     */
-    static class FakeSuccessDriver implements java.sql.Driver {
-
-        static final String URL = "jdbc:athenztest://fake/db";
-
-        @Override
-        public boolean acceptsURL(String url) {
-            return URL.equals(url);
-        }
-
-        @Override
-        public java.sql.Connection connect(String url, java.util.Properties info) {
-            return (java.sql.Connection) java.lang.reflect.Proxy.newProxyInstance(
-                    java.sql.Connection.class.getClassLoader(),
-                    new Class<?>[] { java.sql.Connection.class },
-                    (proxy, method, args) -> {
-                        switch (method.getName()) {
-                            case "isClosed":
-                                return false;
-                            case "equals":
-                                return proxy == (args != null && args.length > 0 ? args[0] : null);
-                            case "hashCode":
-                                return System.identityHashCode(proxy);
-                            case "toString":
-                                return "FakeConnection";
-                            default:
-                                return null;
-                        }
-                    });
-        }
-
-        @Override
-        public int getMajorVersion() {
-            return 1;
-        }
-
-        @Override
-        public int getMinorVersion() {
-            return 0;
-        }
-
-        @Override
-        public boolean jdbcCompliant() {
-            return false;
-        }
-
-        @Override
-        public java.util.logging.Logger getParentLogger() {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public java.sql.DriverPropertyInfo[] getPropertyInfo(String url, java.util.Properties info) {
-            return new java.sql.DriverPropertyInfo[0];
-        }
-    }
-
     @Test
-    public void testVerifyConnectionRealImplementationSucceeds() throws java.sql.SQLException {
+    public void testVerifyConnectionRealImplementationSucceeds() throws SQLException {
 
         // registers a minimal fake driver so the real (non-overridden) verifyConnection()
         // can exercise its success path (a real, non-null Connection) without a live database
 
         FakeSuccessDriver driver = new FakeSuccessDriver();
-        java.sql.DriverManager.registerDriver(driver);
+        DriverManager.registerDriver(driver);
         try {
             AWSObjectStoreFactory factory = new AWSObjectStoreFactory();
             assertTrue(factory.verifyConnection(FakeSuccessDriver.URL, "user", "token"));
         } finally {
-            java.sql.DriverManager.deregisterDriver(driver);
+            DriverManager.deregisterDriver(driver);
         }
     }
